Fix double accounting of skipped objects during copying
When an object is added to skipped_blocks_map_ due to thread
loosing the race on installing forwarding pointer during copying,
we add the size of the object to num_bytes_allocated_. Later,
when an object is allocated out of skipped_blocks_map_,
num_bytes_allocated_ is not adjusted accordingly. So, when this
object gets allocated (or gets added to skipped_blocks_map_
again due to loosing the race), num_bytes_allocated_ goes off.

Bug: 74763563
Test: make test-art
Change-Id: I197b94b317499281e379733d53ad19794dbe772a
Signed-off-by: mydongistiny <jaysonedson@gmail.com>